Prevent creation of empty NuGet.Config file
As Chocolatey, we never want to create the NuGet.Config file, as this is
not used by Chocolatey at all. As such, rather than GetOrCreateDocument,
if the filepath doesn't exist, simply return the default empty
XDocument, which is exactly what would have been returned after creating
the file.

A few tests have to be skipped, as they rely on the config being
automatically created.

Based on this commit:
chocolatey/nuget-chocolatey@a211f19

Co-Authored-By: Gary Ewan Park <1271146+gep13@users.noreply.github.com>