Add IHttpSource and allow overriding of HttpSource
This pulls an interface out of HttpSource called IHttpSource, and
switches usages to use the interface. Additionally, it adds a method
to HttpSourceResource to allow the IHttpSource to be switched out if
needed, so alternative implementations can be used.

Based on these commits:
chocolatey/nuget-chocolatey@19986e7
chocolatey/nuget-chocolatey@ee5f731e

Co-Authored-By: Rob Reynolds <63502+ferventcoder@users.noreply.github.com>